+++ hutool-extra/src/main/java/cn/hutool/extra/pinyin/engine/houbbpinyin/package-info.java
@@ -17,4 +17,4 @@
  *
  * @author looly
  */
-package cn.hutool.extra.pinyin.engine.pinyin;
+package cn.hutool.extra.pinyin.engine.houbbpinyin;