+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * 【http   】     SoapClient增加针对不同协议的头信息（pr#305@Gitee）
 * 【http   】     HttpRequest支持307、308状态码识别（issue#1504@Github）
 * 【core   】     CharUtil.isBlankChar增加\u0000判断（pr#1505@Github）
+* 【extra  】     添加Houbb Pinyin支持（pr#1506@Github）
 
 ### Bug修复
 * 【core   】     修复Validator.isUrl()传空返回true（issue#I3ETTY@Gitee）